extracted inner public classes
